Basic tests with locked games
At this point I'm much more inclined turning this all into a server-side approach with login, permissions, ...
It'll be a pain to basically redo the entire website from scratch (or the logic at least) but it will be better in the long run.
The 5MB limit of the browser's local storage is not helping.